Continue create Auth features and Update Product features

+++ src/Application/NorthwindApi.Application/Features/Auth/Commands/DeleteUserCommand.cs
@@ -7,13 +7,12 @@ using NorthwindApi.Domain.Entities;
 
 namespace NorthwindApi.Application.Features.Auth.Commands;
 
-public record DeleteUserCommand(int UserId) : ICommand<ApiResponse>
-{
-    public int UserId { get; init; } = UserId;
-}
+public record DeleteUserCommand(int UserId) : ICommand<ApiResponse>;
 
 internal class DeleteUserCommandHandler(
     ICrudService<User, int> crudService,
+    ICrudService<UserToken, int> userTokenCrudService,
+    IRepository<UserToken, int> userTokenRepository,
     IUnitOfWork unitOfWork
 ) : ICommandHandler<DeleteUserCommand, ApiResponse>
 {
@@ -26,6 +25,17 @@ internal class DeleteUserCommandHandler(
                 return new ApiResponse(StatusCodes.Status404NotFound, "User not found!!!");
 
             await crudService.DeleteAsync(user, cancellationToken);
+            var userTokens = await userTokenRepository
+                .ToListAsync(
+                userTokenRepository.GetQueryableSet()
+                    .Where(x => x.UserId == command.UserId));
+            if (userTokens.Count >= 0)
+            {
+                foreach (var userToken in userTokens)
+                {
+                    await userTokenCrudService.DeleteAsync(userToken, cancellationToken);
+                }
+            }
             await unitOfWork.CommitTransactionAsync(cancellationToken);
             return new ApiResponse(StatusCodes.Status200OK, "Delete User successfully");
         }
